fix: correct pipeline error behavior
The behavior of the mock was different than the one in `go-redis`.
In the real implementation the error of the last command is returned and all commands are executed (see https://github.com/redis/go-redis/blob/f752b9a9d5cc158381c2ffe5b13c531037426b39/redis.go#L501C22-L524)

In redismock the execution stopped on the first error, which is wrong.